+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.2.0 / 2019-06-25
+==================
+
+* Filter by pending requests and resourceType
+* Assert by pending requests and resourceType
+* Browser.elementHtml and assertions
+* Types export for TypeScript improved
+
 2.1.3/ 2018-06-21
 =================